moved testSparseCheckout to GitAPITestUpdateCliGit
