+++ src/test/java/org/jenkinsci/plugins/gitclient/GitAPITestUpdateCliGit.java
@@ -6,6 +6,8 @@ import org.jvnet.hudson.test.Issue;
 
 import java.io.File;
 import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.UUID;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -186,6 +188,61 @@ public abstract class GitAPITestUpdateCliGit extends GitAPITestUpdate{
 
         checkoutTimeout = 1 + random.nextInt(60 * 24);
         w.git.checkout().branch(DEFAULT_MIRROR_BRANCH_NAME).ref("origin/" + DEFAULT_MIRROR_BRANCH_NAME).timeout(checkoutTimeout).deleteBranchIfExist(true).execute();
+    }
+
+    @Test
+    public void testSparseCheckout() throws Exception {
+        /* Sparse checkout was added in git 1.7.0, but the checkout -f syntax
+         * required by the plugin implementation does not work in git 1.7.1.
+         */
+        if (!w.cgit().isAtLeastVersion(1, 7, 9, 0)) {
+            return;
+        }
+        // Create a repo for cloning purpose
+        w.init();
+        w.commitEmpty("init");
+        assertTrue("mkdir dir1 failed", w.file("dir1").mkdir());
+        w.touch("dir1/file1");
+        assertTrue("mkdir dir2 failed", w.file("dir2").mkdir());
+        w.touch("dir2/file2");
+        assertTrue("mkdir dir3 failed", w.file("dir3").mkdir());
+        w.touch("dir3/file3");
+        w.git.add("dir1/file1");
+        w.git.add("dir2/file2");
+        w.git.add("dir3/file3");
+        w.git.commit("commit");
+
+        // Clone it
+        WorkingArea workingArea = new WorkingArea();
+        workingArea.git.clone_().url(w.repoPath()).execute();
+
+        checkoutTimeout = 1 + random.nextInt(60 * 24);
+        workingArea.git.checkout().ref(defaultRemoteBranchName).branch(defaultBranchName).deleteBranchIfExist(true).sparseCheckoutPaths(Collections.singletonList("dir1")).timeout(checkoutTimeout).execute();
+        assertTrue(workingArea.exists("dir1"));
+        assertFalse(workingArea.exists("dir2"));
+        assertFalse(workingArea.exists("dir3"));
+
+        workingArea.git.checkout().ref(defaultRemoteBranchName).branch(defaultBranchName).deleteBranchIfExist(true).sparseCheckoutPaths(Collections.singletonList("dir2")).timeout(checkoutTimeout).execute();
+        assertFalse(workingArea.exists("dir1"));
+        assertTrue(workingArea.exists("dir2"));
+        assertFalse(workingArea.exists("dir3"));
+
+        workingArea.git.checkout().ref(defaultRemoteBranchName).branch(defaultBranchName).deleteBranchIfExist(true).sparseCheckoutPaths(Arrays.asList("dir1", "dir2")).timeout(checkoutTimeout).execute();
+        assertTrue(workingArea.exists("dir1"));
+        assertTrue(workingArea.exists("dir2"));
+        assertFalse(workingArea.exists("dir3"));
+
+        workingArea.git.checkout().ref(defaultRemoteBranchName).branch(defaultBranchName).deleteBranchIfExist(true).sparseCheckoutPaths(Collections.emptyList()).timeout(checkoutTimeout).execute();
+        assertTrue(workingArea.exists("dir1"));
+        assertTrue(workingArea.exists("dir2"));
+        assertTrue(workingArea.exists("dir3"));
+
+        workingArea.git.checkout().ref(defaultRemoteBranchName).branch(defaultBranchName).deleteBranchIfExist(true).sparseCheckoutPaths(null)
+                .timeout(checkoutTimeout)
+                .execute();
+        assertTrue(workingArea.exists("dir1"));
+        assertTrue(workingArea.exists("dir2"));
+        assertTrue(workingArea.exists("dir3"));
     }
 
 }
